chore: finish current stage of legacy implementation
